requirements.txt
chess
pandas
python-telegram-bot
cairosvg
